+++ modules/visual-tests/src/layout/ReviewShot.tsx
@@ -2,12 +2,11 @@ import React, { useEffect, useRef, useState } from 'react';
 import { LinksFunction } from '@remix-run/node';
 import { Form, Link, useTransition } from '@remix-run/react';
 import TagPill, { TagPillProps } from '@reykjavik/hanna-react/TagPill';
+import { capitalize } from '@reykjavik/hanna-utils';
 
 import { Changeset } from '../utils/tests.server';
 
 import styles from './ReviewShot.css';
-
-const capitalize = (str: string) => str[0]?.toUpperCase() + str.slice(1);
 
 export type Mode = 'difference' | 'actual' | 'expected';